Refactor package globals into container struct (#1)
This PR attempts to eliminate globals in the Goose package, instead introducing goose.Client as a container for the data that was previously stored in globals. Existing methods are refactored to use a single global instance of this struct.

One of the advantages of this approach is the ability to define separate database tables to track migrations separately for creating tables and populating data.

There is also cleanup of typos and dead code.